+++ platforms/core-execution/build-cache-example-client/src/main/java/org/gradle/caching/example/ExampleBuildCacheClient.java
@@ -62,10 +62,10 @@ import org.gradle.internal.concurrent.ExecutorFactory;
 import org.gradle.internal.file.FileAccessTimeJournal;
 import org.gradle.internal.file.FileAccessTracker;
 import org.gradle.internal.file.FileMetadataAccessor;
-import org.gradle.internal.file.ModificationTimeFileAccessTimeJournal;
 import org.gradle.internal.file.StatStatistics;
 import org.gradle.internal.file.TreeType;
 import org.gradle.internal.file.impl.SingleDepthFileAccessTracker;
+import org.gradle.internal.file.nio.ModificationTimeFileAccessTimeJournal;
 import org.gradle.internal.file.nio.NioFileMetadataAccessor;
 import org.gradle.internal.file.nio.PosixJdk7FilePermissionHandler;
 import org.gradle.internal.hash.DefaultFileHasher;
