+++ platforms/core-execution/persistent-cache/src/test/groovy/org/gradle/cache/internal/LeastRecentlyUsedCacheCleanupTest.groovy
@@ -18,7 +18,7 @@ package org.gradle.cache.internal
 
 import org.gradle.cache.CleanableStore
 import org.gradle.cache.CleanupProgressMonitor
-import org.gradle.internal.file.ModificationTimeFileAccessTimeJournal
+import org.gradle.internal.file.nio.ModificationTimeFileAccessTimeJournal
 import org.gradle.internal.time.TimestampSuppliers
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
 import org.junit.Rule
