+++ platforms/core-runtime/files/src/main/java/org/gradle/internal/file/nio/ModificationTimeFileAccessTimeJournal.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.file;
+package org.gradle.internal.file.nio;
 
+import org.gradle.internal.file.FileAccessTimeJournal;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -24,6 +25,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.attribute.FileTime;
 
+@SuppressWarnings("Since15")
 public class ModificationTimeFileAccessTimeJournal implements FileAccessTimeJournal {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ModificationTimeFileAccessTimeJournal.class);
